pyproject.toml
[project]
name = "deepsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langchain-tavily>=0.1.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
    "streamlit>=1.28.0",
]
